% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pkg.R
\docType{package}
\name{BelgiumPopulation-package}
\alias{BelgiumPopulation-package}
\alias{ETLUtils}
\alias{ETLUtils-package}
\title{Belgium population statistics available as open data}
\description{
More information: http://statbel.fgov.be/nl/statistieken/opendata/home
}
\examples{
data(tf_births)
}

